requirements-desktop-win.txt
pystray>=0.19.0
pywebview>=5.0
Pillow>=10.0
pyinstaller>=6.0
